[ICU-12611] Fix some issues found through static analysis. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/39151 Former-commit-id: 2aa92371843194881d3d72e754001907c89023ce

+++ icu4j/main/classes/collate/src/com/ibm/icu/impl/coll/CollationFastLatinBuilder.java
@@ -45,7 +45,7 @@ final class CollationFastLatinBuilder {
         if (limit == 0) { return ~0; }
         int start = 0;
         for (;;) {
-            int i = (start + limit) / 2;
+            int i = (int)(((long)start + (long)limit) / 2);
             int cmp = compareInt64AsUnsigned(ce, list[i]);
             if (cmp == 0) {
                 return i;
